n0_error
Attribute Macro add_meta

#[add_meta]

Attribute macro that adds a meta: ::n0_error::Meta field to a struct or to all named-field variants of an enum. Does nothing else.

If the struct or an enum variant is currently a unit kind, it will be converted to a named-field kind.

Tuple structs or variants are not supported.