n0_error
Struct AnyError

pub struct AnyError(/* private fields */);

Type-erased error that can wrap a StackError or any std::error::Error.

StackErrors have a blanket impl to convert into AnyError, while preserving their call-site location info.

Errors that implement std::error::Error but not StackError canโt convert to AnyError automatically. Use either AnyError::from_std or crate::StdResultExt::anyerr to convert std errors into AnyError.

This is necessary unfortunately because if we had a blanket conversion from std errors to AnyError, this blanket conversion would also apply to StackErrors, thus losing their call-site location info in the conversion.

Implementations

impl AnyError

pub fn from_std(err: impl Error + Send + Sync + 'static) -> Self

Creates an AnyError from a std error.

This captures call-site metadata.

pub fn from_anyhow(err: Error) -> Self

Creates an AnyError from a anyhow::Error.

pub fn from_std_box(err: Box<dyn Error + Send + Sync + 'static>) -> Self

Creates an AnyError from a bosed std error.

This captures call-site metadata.

pub fn from_display(s: impl Display) -> Self

Creates an AnyError from any Display value by formatting it into a message.

pub fn from_string(message: String) -> Self

Creates an AnyError from a message string.

pub fn from_stack(err: impl StackError + 'static) -> Self

Creates an AnyError from a StackError.

This preserves the errorโs call-site metadata.

Equivalent to AnyError::from(err).

pub fn from_stack_box(err: Box<dyn StackError>) -> Self

Creates an AnyError from a boxed StackError.

pub fn context(self, context: impl Display) -> AnyError

Adds additional context on top of this error.

pub fn into_boxed_dyn_error(self) -> Box<dyn Error + Send + Sync + 'static>

Converts into boxed std error.

pub fn downcast_ref<T: Error + 'static>(&self) -> Option<&T>

Downcast this error object by reference.

pub fn downcast<T: Error + 'static>(self) -> Option<T>

Downcast this error object by reference.