n0_error
Trait StdResultExt

pub trait StdResultExt<T, E> {
    // Required methods
    fn std_context(self, context: impl Display) -> Result<T, AnyError>;
    fn with_std_context<F, C>(self, context: F) -> Result<T, AnyError>
       where F: FnOnce(&E) -> C,
             C: Display;
    fn e(self) -> Result<T, AnyError>;
}

Provides extension methods to add context to std errors.

Required Methods

fn std_context(self, context: impl Display) -> Result<T, AnyError>

Wraps the result’s error value with additional context.

fn with_std_context<F, C>(self, context: F) -> Result<T, AnyError>
where F: FnOnce(&E) -> C, C: Display,

Wraps the result’s error value with lazily-evaluated additional context.

The context closure is only invoked if an error occurs.

fn e(self) -> Result<T, AnyError>

Converts the result’s error into AnyError.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.